+++ src/App.ts
@@ -1,5 +1,3 @@
-/// <reference path="./../typings/typings.d.ts" />
-
 import AbstractContext from "./AbstractContext";
 import Context from "./Context";
 import * as Document from "./Document";