+++ src/Committer.ts
@@ -1,5 +1,3 @@
-/// <reference path="./../typings/typings.d.ts" />
-
 interface Committer<E> {
 	commit( object:E ):Promise<any>;
 }