+++ src/Errors/IllegalStateError.spec.ts
@@ -1,5 +1,3 @@
-/// <reference path="../../typings/typings.d.ts" />
-
 import IllegalStateError from "./IllegalStateError";
 import AbstractError from "./AbstractError";
 import * as Utils from "./../Utils";