+++ src/JSONLDConverter.ts
@@ -1,5 +1,3 @@
-/// <reference path="./../typings/typings.d.ts" />
-
 import * as Errors from "./Errors";
 import * as ObjectSchema from "./ObjectSchema";
 import * as NS from "./NS";