+++ src/LDP/AccessPoint.ts
@@ -1,5 +1,3 @@
-/// <reference path="../../typings/typings.d.ts" />
-
 import * as Container from "./Container";
 import * as NS from "./../NS";
 import * as ObjectSchema from "./../ObjectSchema";